backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-magic-bin>=0.4.14;sys_platform=='win32'
python-magic>=0.4.27;sys_platform!='win32'
